Refactor tests

+++ lib/include/syscalls.h
@@ -15,7 +15,7 @@ int open_frame_buffer(int);
 
 unsigned char * read_frame_buffer(size_t buffer_size, off_t offset, off_t pa_offset);
 
-int write_to_frame_buffer(int *write_buffer, size_t num_bytes);
+int write_to_frame_buffer(unsigned short *write_buffer, size_t num_bytes);
 
 void exit_graphics();// ioctl
