+++ lib/src/syscalls.c
@@ -18,7 +18,7 @@ int open_frame_buffer(int options) {
     return open("/dev/fb0", options);
 }
 
-int write_to_frame_buffer(int *write_buffer, size_t num_bytes) {
+int write_to_frame_buffer(unsigned short *write_buffer, size_t num_bytes) {
 
     int filedesc = open_frame_buffer(O_WRONLY | O_APPEND);
